requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Git Operations
gitpython==3.1.41

# Code Parsing (多语言支持)
tree-sitter==0.21.0
tree-sitter-java==0.21.0
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
tree-sitter-go==0.21.0
tree-sitter-cpp==0.21.0
tree-sitter-rust==0.21.0
javalang==0.13.0
libcst==1.1.0  # Python AST

# AI Models
openai==1.10.0
anthropic==0.18.0
langchain==0.1.4
langchain-community==0.0.16
tiktoken==0.6.0

# Additional AI Models Support
google-generativeai==0.3.2  # Google Gemini
dashscope==1.14.0  # 阿里通义千问

# Code Formatting & Analysis
black==24.1.1
isort==5.13.2
autopep8==2.0.4
pylint==3.0.3
mypy==1.8.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
rich==13.7.0  # 终端美化
click==8.1.7  # CLI
httpx==0.26.0
aiofiles==23.2.1
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Processing
zipfile36==0.1.3
pathspec==0.12.1

